removed callbacks and layout from main
